[xbase][lib] Renamed IterableExtension#elementsToString(..) to #join(..) [xtend][xbase] Regenerated purexbase and xtend code with latest version

+++ examples/org.eclipse.xtext.xtext.ui.examples/contents/org.eclipse.xtext.example.xbase.tutorial/src-gen/Xbase01_Literals.java
@@ -9,9 +9,9 @@ public class Xbase01_Literals {
 		{/*null*/;/*"Hello World"*/;/*42*/;/*true*/;/*false*/;/*java.lang.String.class*/;
 		  CollectionLiterals.<Integer>newArrayList(((Integer)3), ((Integer)5), ((Integer)7));
 		  CollectionLiterals.<String>newHashSet("foo", "bar", "baz");
-		  Pair<Integer,String> _operator_mappedTo = ObjectExtensions.<Integer, String>operator_mappedTo(1, "one");
-		  Pair<Integer,String> _operator_mappedTo_1 = ObjectExtensions.<Integer, String>operator_mappedTo(2, "two");
-		  Pair<Integer,String> _operator_mappedTo_2 = ObjectExtensions.<Integer, String>operator_mappedTo(3, "three");
+		  Pair<Integer,String> _operator_mappedTo = ObjectExtensions.<Integer, String>operator_mappedTo(((Integer)1), "one");
+		  Pair<Integer,String> _operator_mappedTo_1 = ObjectExtensions.<Integer, String>operator_mappedTo(((Integer)2), "two");
+		  Pair<Integer,String> _operator_mappedTo_2 = ObjectExtensions.<Integer, String>operator_mappedTo(((Integer)3), "three");
 		  CollectionLiterals.<Integer, String>newHashMap(_operator_mappedTo, _operator_mappedTo_1, _operator_mappedTo_2);
 		}
 	}
