+++ examples/org.eclipse.xtext.xtext.ui.examples/contents/org.eclipse.xtext.example.xbase.tutorial/src-gen/Xbase06_If.java
@@ -6,17 +6,17 @@ public class Xbase06_If {
 		
 		{
 		  int _length = "foo".length();
-		  boolean _operator_greaterThan = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length), 3);
+		  boolean _operator_greaterThan = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length), ((Integer)3));
 		  if (_operator_greaterThan) {/*"never true!"*/;
 		  } else {/*"always true!"*/;
 		  }
 		  int _length_1 = "foo".length();
-		  boolean _operator_greaterThan_1 = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length_1), 3);
+		  boolean _operator_greaterThan_1 = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length_1), ((Integer)3));
 		  if (_operator_greaterThan_1) {/*"never true!"*/;
 		  }
 		  boolean _xifexpression = false;
 		  int _length_2 = "foo".length();
-		  boolean _operator_greaterThan_2 = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length_2), 3);
+		  boolean _operator_greaterThan_2 = ComparableExtensions.<Integer>operator_greaterThan(((Comparable<? super Integer>)_length_2), ((Integer)3));
 		  if (_operator_greaterThan_2) {
 		    _xifexpression = false;
 		  } else {
